+++ tests/utils/test_timeframe.py
@@ -66,6 +66,30 @@ def test_round_down_t10_minutes():
     ) == datetime(2023, 6, 6, 12, 0, 0)
 
 
+def test_round_down_t10_minutes_two():
+    assert round_down_timestamp(
+        datetime(2023, 6, 6, 12, 19, 43, 3857), timedelta(minutes=10)
+    ) == datetime(2023, 6, 6, 12, 10, 0)
+
+
+def test_round_down_h1():
+    assert round_down_timestamp(
+        datetime(2023, 6, 6, 12, 9, 43, 3857), timedelta(hours=1)
+    ) == datetime(2023, 6, 6, 12, 0, 0)
+
+
+def test_round_down_h1_same():
+    assert round_down_timestamp(datetime(2023, 6, 6, 12, 0, 0, 0), timedelta(hours=1)) == datetime(
+        2023, 6, 6, 12, 0, 0
+    )
+
+
+def test_round_down_d1():
+    assert round_down_timestamp(
+        datetime(2023, 6, 6, 12, 9, 43, 3857), timedelta(days=1)
+    ) == datetime(2023, 6, 6, 0, 0, 0)
+
+
 def test_timedelta_to_str_seconds():
     assert timedelta_to_str(timedelta(seconds=30)) == "S30"